requirements.txt
certifi
charset-normalizer
click
colorama
Flask
greenlet
idna
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
pydantic
python-dotenv
requests
SQLAlchemy==1.4.48
typing_extensions
urllib3
Werkzeug
